Separate data_handling functions from main
